+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.5] - 2025-12-31
+
+### Fixed
+
+- Fixed port conflict on puppeteer auto-install restart: WebSocket server now properly stops before process restart
+
+## [1.1.4] - 2025-12-31
+
+### Fixed
+
+- Fixed puppeteer auto-install: Process now restarts after installation to pick up newly installed module (ESM import cache issue)
+
 ## [1.1.3] - 2025-12-31
 
 ### Added